IMM: 2PBE slave wait for prepare from primary in PRT-create & PRT-update [#623]
Risk for discrepancy between 2PBE slave and primary is reduced further
by having the slave wait for ok_to_prepare trigger from primary for
the weak-ccb associated with the PRT operations. The 2PBE primary will
also not perform the PRT operation if it does not get ok reply from slave
on the prepare call, This follows the same pattern as existed earlier
(in 4.4 2PBE) for regular CCBs and for PRO deletes.

The ok_to_prepare call is a synchronous admin-opertion invoked from primary
to slave. At primary it blocks the main thread waiting for reply. At slave
it is received and executed by the slave's runtime thread, where the slave's
main thread should have received the data and be blocked waiting for this
ok_to_pripare.

The risk for discrepancy in the sqlite commit to the two pbe files
should be minimal and only possible by crash or SIGKILL of one
or both PBEs during sqlite commit processing. A SIGTERM on the PBEs should
not be able to cause a discrepancy.