amfd: Respond with TRY_AGAIN to admin ops during a controller switch-over [#821]
It has been found that an admin operation (other than controller switch-over)
can be started after a controller switch-over has been initiated.

The result of this can be that the admin operation response gets lost.
The reason for it to get lost is that an IMM admin response has to be
responded on the same node where it was initiated.

The change is to disallow new admin operations when controller
switch-over is in progress. This will result in TRYAGAIN to the issuing program
which should stop once the AMF implementer has connected and the controller
switch-over has finished.

credit: Hans Feldt