amfd: update RT objects during node-failover of active controller [#494].
Run time objects and attributes are not updated/deleted/created
when node-failover gots escalated for active controller and standby
controller took the active role.

Activities related to update of runtime objects and certain
attribute to IMM are given low priotiy and are pushed in Job queue by AMF.
These jobs are completed when AMF is not busy in any other high priority
activity. When node-failover is escalated, AMFD sends reboot message to
AMFND to reboot the node. In case node-failover is escalated for active
controller, it will send reboot message to AMFND which will reboot the
controller. In such a case, some IMM related activites in JOB queue will
remian uncompleted. All such activites should be compleleted before
rebooting the active controller when node-failover is escalated for it.

Patch will try to finish as many IMM related jobs as possbile just
after sending reboot message to AMFND when node-failover is escalated
for active controller.