base: add secutil for sending, receiving over a UNIX socket [#554]
This patch add some infrastructure that can be used by other services
to securely get credentials for a client.

When initialized on the server side a tiny server thread is created
that listen to a named connection oriented UNIX socket. For each incoming
connection request it calls a user specified callback and handover
file descriptor and peer credentials.

A typical use case is for a library to create the MDS socket and then
use the client side function to send the MDS address as a data messages.
The client credentials can be stored on the server side and verified
when needed.

Note each received connection less message needs to be checked with the
original received MDS address so no spoofing (handle hijacking) is
happening.