clm: Ensure SaNameT strings are NUL-terminated [#1056]
CLM uses the SaNameT structure for output parameters (in various data
structures), but it did not guarantee that the string inside the SaNameT
structure was NUL-terminated. This meant that these SaNameT structures could not
safely be read using the saAisNameBorrow() function.

Code has been added to the clm agent to ensure proper NUL termination in SaNameT
structures.